requirements.txt
beautifulsoup4==4.12.3
folium==0.16.0
pandas==2.2.1
python-dotenv==1.1.1
